pyproject.toml
[project]
name = "meowmx"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "psycopg>=3.2.9",
    "python-event-sourcery==0.4.0a15",
    "sqlalchemy>=2.0.43",
]

[dependency-groups]
dev = [
    "coolname>=2.2.0",
    "mypy>=1.17.0",
    "pytest>=8.4.1",
    "ruff>=0.12.4",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.mypy]
check_untyped_defs = true
